requirements.txt
flask
torch
torchvision
faiss-cpu
numpy
Pillow
